requirements.txt
Django==5.1.6
Pillow>=10.0,<12.0
